third_party/WebKit/LayoutTests/media/controls/accessibility-timeline.html

<!DOCTYPE html>
<title>Media Controls: timeline accessibility tests</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../media-controls.js"></script>
<video controls></video>
<script>
async_test(t => {
  assert_true(internals.runtimeFlags.accessibilityObjectModelEnabled);

  var video = document.querySelector('video');
  video.src = '../content/test.ogv';

  var timeline = timelineElement(video);
  assert_not_equals(timeline, null);

  video.play().then(t.step_func(() => {
    assert_equals(timeline.getAttribute('aria-valuetext'), '0:00');

    // Let the video play a bit before checking the aria-valuetext again.
    setTimeout(t.step_func_done(() => {
      assert_equals(timeline.getAttribute('aria-valuetext'), '0:01');
    }), 1500);
  }));
});
</script>
